Proper markdown comments.
Apparently they also have parenthesis.  Fine.

The syntactic load of that is sort of horrific, but if the goal of
"the data is in markdown documents that are also being parsed by other things"
appeals to you, then this is where you have to live.

(Or we'd have to make a different convention, like using h4 tags, or something.
But that seems more opinionated, and wouldn't always fit with other tools.)