Further improve readme.
Work on first impression text.

Add ToC links.  (Important to make sure people see that there's
description of both the format and the motivations and some code;
don't want them to get burned out on the boring motivation stuff!)

Added short heading for each major feature category of the library.

That includes describing the dir-like indexing feature.

Put some example code in plain view in the readme.

Linked to other examples of use in the wild, which are probably good
stuff for riffing from.