refrain from making double linebreaks when patching with content that ends in a linebreak.
The vagaries of split behavior were not helpful here.

Unconditionally sheering the last entry of the line slice would also be
incorrect, because we generally have a tolerant view of data that
doesn't end in line breaks.  (We just have no choice but to normalize
it to ending in a line break due to the nature of markdown codeblocks.)